Add @ytt:lib to load and render ytt top-level templates
@ytt:lib can be loaded from within starlark and it provides a `lib`
modules with a `load()` function. it takes:

  - A string with the location of the module to load
  - An optional `recursive` keyword argument, defaults to true

It returns a loadable library which accepts the `eval()` method. It
takes a single optional argument with the values to override in the
evaluated library

After evaluation, the returned object has an `output()` method to fetch
the result from the library evaluation. It takes an optional string
argument for the file name to get the results from. If empty, it will
return the concatenated results for all files. To get results as string,
there is an equivalent functtion called `text()`.

This is quite equivalent as running ytt from within the template and
embedding the output in it.