Add @ytt:lib to load and render ytt top-level templates
@ytt:lib can be loaded from within starlark and it provides a `lib`
modules with a `lib.eval()` function that load and evaluates a library.
It takes:

  - A string with the location of the module to load. It can start with
    `@` in which case it is assumed that the library comes from
    `_ytt_lib/`, and the library is internal
  - An optional value object to override current data/values

After evaluation, the returned object has `output()` and `text()`
methods to fetch the result from the library evaluation. If called
without argument, it returns the complete evaluation. Else, it takes a
file name and returns the evaluation results for that file. `text()`
always returns a string in YAML form. `output()` returns the YAML data
structure.