Remove Broken Schema Definition Link (#246)
Until docs are in more mature state, we should not provide a link with this error. Currently docs live in the design docs repo (https://github.com/k14s/design-docs/blob/develop/ytt/), but they are not up to date with the proposal and may cause confusion. A link should be added back when schema docs are better defined.