WIP: half done
So basically this is half done. If the file name is not .txt it
displays filename
Not sure if test coverage is enought
Missing the part when the file extension is txt. we need to understand
what do to in that place. Maybe we should talk to John to see if he has
any idea