adding check for if issue closed timestamp != comment created time stamp
Adding a check to prevent the carvel-triage label from being added when an issue is closed with a comment.

Issues being closed with a comment will have the same time stamp for both the issue closing/comment, so this check makes sure the time stamps are different (i.e. the comment happened after the issue was closed so the carvel-triage label should be reapplied).