Refactor:
- Make it clear that Node is a sealed interface. And why it is the case.

A sealed interface: can only be discussed in the
context of having multiple packages. A sealed interface is an interface
with unexported methods. This means users outside the package is unable
to create types that fulfil the interface.

In YTT's context, a Node must only ever have a type that maps
directly to a valid YAML type. Thus having an exhaustive list of
concrete types achieves that.

Authored-by: Dennis Leon <leonde@vmware.com>