Refactor: Introduce DocumentSchemaEnvelope
- Add DocumentSchemaEnvelope
-- The word envelope, hopefully communicates that this type is:
a) A wrapper for a Document Schema
b) is used to locate / deliver a DocumentSchema to its intended address
- Remove 'Name' from Schema
-- The original intent was to lay some of the groundwork to allow targetting a
schema for non-ytt usages. i.e. provide a schema to a k8s resources.
When the time comes for that requirement, re-add Name and all other
changes needed to make that possible.

Authored-by: Dennis Leon <leonde@vmware.com>