Remove redundant go fmt/mod-vendor/mod-tidy
This script was running `go fmt` over the `yaml.v2` package as well.
That code is manually vendored and should not be changed, incidentally.

Note that this script runs ./hack/build.sh.
That script runs `go fmt` _excluding_ the `yaml.v2` package.
It also runs the `go mod...` commands removed here.

When Go is upgraded, sometimes this means that `go fmt` produces
different output. It does for Go 1.19.

Fixes: https://github.com/vmware-tanzu/carvel-ytt/actions/runs/3625263105